Refactor agent history alignment to resampling
- Replaced agent history alignment with a new agent history resampler.
- Updated CMakeLists.txt to include new source and test files.
- Removed agent_history_alignment.hpp and agent_history_alignment.cpp.
- Introduced agent_history_resampler.hpp and agent_history_resampler.cpp.
- Updated AgentData to use resampling parameters instead of alignment parameters.
- Modified tests to reflect changes in history resampling logic.
